feat: add logger middlewar
